mesh_clustering_dashboard/requirements.txt
dash==2.11.1
dash-bootstrap-components==1.6.0
plotly==5.15.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
trimesh==3.20.0
open3d==0.17.0
PyYAML==6.0
Flask-Caching